Introduced modular JS structure
