+++ src/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Wbpack Exercise</title>
+    <title>Webpack Exercise</title>
   </head>
   <body>
     <div id="taskContainer">
@@ -12,7 +12,7 @@
         <ul class="description-list"></ul>
         <div class="todoContaiter"></div>
       </form>
-      <button class="button-for-reset"> Clear all completed </button>
+      <button class="button-for-reset">Clear all completed</button>
     </div>
     <script src="../dist/bundle.js"></script>
   </body>
